chore: move the parser to service
